# openapi.yaml
openapi: "3.0.1"
info:
  title: "API Documentation"
  version: "1.0"
paths:
...
  /api/v2/content/news/page:
    get:
      summary: "新闻列表"
      description: "新闻列表"
      parameters:
      - name: "page"
        in: "query"
        description: ""
        required: true
        schema:
          type: "integer"
          nullable: false
          default: "1"
      - name: "size"
        in: "query"
        description: ""
        required: true
        schema:
          type: "integer"
          nullable: false
          default: "10"
      - name: "keyword"
        in: "query"
        description: ""
        required: false
        schema:
          type: "string"
          nullable: false
      - name: "type"
        in: "query"
        description: ""
        required: false
        schema:
          type: "string"
          nullable: false
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "object"
                    properties:
                      records:
                        type: "array"
                        description: "当前页数据"
                        items:
                          type: "object"
                          properties:
                            id:
                              type: "integer"
                              description: "新闻ID（主键，自增）"
                              format: "int64"
                            author:
                              type: "integer"
                              description: "作者（关联 tb_user.id）"
                              format: "int64"
                            title:
                              type: "string"
                              description: "标题"
                            cover:
                              type: "string"
                              description: "封面"
                            type:
                              type: "string"
                              description: "类型"
                            status:
                              type: "integer"
                              description: "状态（1、2）"
                            content:
                              type: "string"
                              description: "正文"
                            viewCount:
                              type: "integer"
                              description: "浏览量"
                            createTime:
                              type: "string"
                              description: "创建时间"
                            updateTime:
                              type: "string"
                              description: "修改时间"
                          description: "新闻实体\n对应数据库表：tb_news"
                      total:
                        type: "integer"
                        description: "总记录数"
                        format: "int64"
                      page:
                        type: "integer"
                        description: "当前页码（从 1 开始）"
                        format: "int64"
                      size:
                        type: "integer"
                        description: "每页大小"
                        format: "int64"
                    description: ""
  /api/v2/content/news/{id}:
    get:
      summary: "新闻详情"
      description: "新闻详情"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        schema:
          type: "integer"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: ""
                  message:
                    type: "string"
                    description: ""
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: "新闻ID（主键，自增）"
                        format: "int64"
                      author:
                        type: "integer"
                        description: "作者（关联 tb_user.id）"
                        format: "int64"
                      title:
                        type: "string"
                        description: "标题"
                      cover:
                        type: "string"
                        description: "封面"
                      type:
                        type: "string"
                        description: "类型"
                      status:
                        type: "integer"
                        description: "状态（1、2）"
                      content:
                        type: "string"
                        description: "正文"
                      viewCount:
                        type: "integer"
                        description: "浏览量"
                      createTime:
                        type: "string"
                        description: "创建时间"
                      updateTime:
                        type: "string"
                        description: "修改时间"
                    description: ""
...